requirements.txt
async-timeout==4.0.3
asyncpg==0.29.0
python-dotenv==1.0.1
tqdm==4.66.2